feat(contacts): manage contacts from threads

Copilot-Session: 369472a4-40f8-4604-81c2-c4942b88f0b6
